Fixed C warning.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@57080 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/run-time/malloc.c
@@ -2079,7 +2079,7 @@ rt_private union overhead *add_core(size_t nbytes, int type)
 {
 	RT_GET_CONTEXT	
 	union overhead *oldbrk; /* Initialized with `failed' value. */
-	int mod;					/* Remainder for padding */
+	size_t mod;					/* Remainder for padding */
 	size_t asked = nbytes;	/* Bytes requested */
 
 		/* We want at least 'nbytes' bytes for use, so we must add the overhead
